// src/modules/solanafm/solanafm.schema.ts
export interface SolanaFMTransactionsRequest {
    accountHash: string;
    actions?: string;
    utcFrom?: number;
    utcTo?: number;
    inflow?: boolean;
    outflow?: boolean;
    mint?: string;
    mints?: string;
    amountFrom?: number;
    amountTo?: number;
    program?: string;
    programs?: string;
    limit?: number;
    page?: number;
}

export interface SolanaFMError {
    message: string;
    code?: number;
    details?: any;
}

export type TokenType = 'CompressedNonFungible' | 'Fungible' | 'Legacy' | 'NonFungible';
// ...
export interface OwnerTokenAccountsRequest {
    accountHash: string;
    tokenType?: TokenType;
}
// ...
export interface TokenInfoRequest {
    tokenHash: string;
}

export interface TokenSupplyResponse {
    circulatingSupply: number;
    tokenWithheldAmount: number | null;
    userTotalWithheldAmount: number;
    totalWithheldAmount: number;
    realCirculatingSupply: number;
    decimals: number;
}

export interface TokenSupplyRequest {
    tokenHash: string;
}
// ...
export interface TransferRequest {
    hash: string;
}